Make SliderPath mutable
The previous reasoning to make it immutable was for performance reasons - to reduce allocations. This feels really bad from the PoV of the editor, where it has to reconstruct the whole path + segments (in the future) + control points for each change.
Going forward, mutable paths will allow us to add/remove control points and segments, with appropriate structures in the editor to support the path/segment/point hierarchy.