Fix JSON deserialising of PlaylistItem losing data
This is a really weird one and I don't know why it happens. `beatmap_id`
is deserialised first which populates the `Beatmap` backing value, and
then `beatmap` is deserialised where it seems like JSON.NET is calling
the getter of `apiBeatmap` and using that value to determine if
deserialisation should continue.

For reasons I do not know at this stage, JSON.NET believes
deserialisation shouldn't continue for `apiBeatmap`. And it seems like
`APIBeatmap.Equals()` isn't being called to determine this. It is likely
just that this value is non-null which is the determining factor.

There are two fixes here that work together:
1. Remove the setter of `onlineBeatmapId` so `beatmap_id` can't be
   deserialised into it.
2. Make `apiBeatmap` return a `null` value rather than constructing a
   value on the fly.