Refactor `Background` to allow overriding its sprite
We have a pretty strict relationship between `Background` and its
contained `Sprite`, and `BackgroundScreen` and its contained
`Background`.
Though it's not entirely obvious at an initial glance, check out for
example `BackgroundScreenBeatmap` or `OnlinePlayBackgroundScreen`.

What I'd like to do is create a custom sprite that does an online
beatmap lookup of its own.