Prevent multiplayer events getting lost during Join
Matchmaking is somewhat of a special case where every user is expected
to join pretty instantly. The structures here, and really the server as
a whole, are not built to handle this.

For example...

- Client 1 requests to join room
- Server sends room to Client 1.
- Client 1 begins populating users/etc, fires off update schedules.
- Client 2 joins the room.
- Server sends `UserJoined` to Client 1.
- Client 1 does not handle the join event because the pre-tasks haven't
complete yet, and `Room` is still `null`.

---

This is a very hacky way of handling this. It really should not exist,
but I can't think of an easier way to handle for the time being.